move geo helper to GeoUtils
