+++ app/src/org/commcare/android/util/MediaUtil.java
@@ -9,8 +9,6 @@ import android.view.Display;
 import android.view.WindowManager;
 
 import org.commcare.dalvik.preferences.DeveloperPreferences;
-import org.javarosa.core.model.data.GeoPointData;
-import org.javarosa.core.model.data.UncastData;
 import org.javarosa.core.reference.InvalidReferenceException;
 import org.javarosa.core.reference.ReferenceManager;
 
@@ -233,20 +231,4 @@ public class MediaUtil {
         return customDpiScaleFactor * proportionalAdjustmentFactor;
     }
 
-
-    /**
-     * Pass in a string representing either a GeoPoint or an address and get back a valid
-     * GeoURI that can be passed as an intent argument 
-     */
-    public static String getGeoIntentURI(String rawInput){
-        try {
-            GeoPointData mGeoPointData = new GeoPointData().cast(new UncastData(rawInput));
-            String latitude = Double.toString(mGeoPointData.getValue()[0]);
-            String longitude= Double.toString(mGeoPointData.getValue()[1]);
-            return "geo:" + latitude + "," + longitude + "?q=" + latitude + "," + longitude;
-            
-        } catch(IllegalArgumentException iae){
-            return "geo:0,0?q=" + rawInput;
-        }
-    }
 }
